+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
- -
+**Improvements**
+ - `folders.get_folder()` can now handle folder names that include a leading or trailing "/" character.
 
 v1.8.1 (2023-01-19)
 ----------